refactor creation of suborgas
Every orga is the suborga of some orga. So it's not possible to
just add orgaA to the suborgas of orgaB because it's already suborga
of some orgaC, it has to be moved (not part of prototype I)
Also, as there are no "single" orgas, there is no orga#create route.
New orgas are created as suborgas under the "afeefa"-orga